requirements.txt
numpy>=1.21.0
Pillow>=8.3.0
pandas>=1.3.0
